refactor: move common steps to acceptance tester
